TELEGRAM-RESLLER/requirements.txt
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pycryptodome==3.23.0
requests==2.32.5
urllib3==2.5.0
python-telegram-bot
